Move tabular example envs into imitation.examples
